[go-gen] Separate out GoCommonGenerator
