+++ src/main/scala/temple/generate/service/go/common/GoCommonCommGenerator.scala
@@ -1,0 +1,9 @@
+package temple.generate.service.go.common
+
+import temple.utils.FileUtils
+
+object GoCommonCommGenerator {
+
+  private[go] def generateInit(): String =
+    FileUtils.readResources("go/genFiles/comm_init.go").stripLineEnd
+}
