+++ src/main/scala/temple/generate/service/go/common/GoCommonMainGenerator.scala
@@ -1,0 +1,9 @@
+package temple.generate.service.go.common
+
+import temple.utils.FileUtils
+
+object GoCommonMainGenerator {
+
+  private[go] def generateJsonMiddleware(): String =
+    FileUtils.readResources("go/genFiles/json_middleware.go").stripLineEnd
+}
